Move Commands from application into module
